Signed resource integrated to SAS tokens.
New resource identifier in addition to "b" and "c": "bs" for blob
snapshots.
SAS tokens now have two extra fields in the signature, signed resource and snapshot ID.
Signed resource is the resource identifier on the token.
snapshot ID is the timestamp id associated with a blob snapshot, if
granting access to a specific blob snapshot.